Exam conf moved to MySql

+++ src/main/resources/db/migration/V1__Create_users_table.sql
@@ -1,21 +1,24 @@
 CREATE TABLE `student_groups` (
-  `id`   BIGINT NOT NULL AUTO_INCREMENT,
+  `id`   BIGINT       NOT NULL AUTO_INCREMENT,
   `name` VARCHAR(100) NOT NULL UNIQUE,
   PRIMARY KEY (`id`)
-) DEFAULT CHARSET=utf8;
+)
+  DEFAULT CHARSET = utf8;
 
 CREATE TABLE `users` (
-  `id`          BIGINT NOT NULL AUTO_INCREMENT,
-  `username`    VARCHAR(100) NOT NULL UNIQUE,
-  `password`    VARCHAR(100) NOT NULL,
-  `first_name`  VARCHAR(255) NOT NULL,
-  `last_name`   VARCHAR(255) NOT NULL,
-  `email`       VARCHAR(255) NOT NULL,
-  `user_type`   INT(11) NOT NULL,
-  `group_id`    BIGINT NULL,
-  `access_key`  VARCHAR(100) NOT NULL,
+  `id`         BIGINT       NOT NULL AUTO_INCREMENT,
+  `username`   VARCHAR(100) NOT NULL UNIQUE,
+  `password`   VARCHAR(100) NOT NULL,
+  `first_name` VARCHAR(255) NOT NULL,
+  `last_name`  VARCHAR(255) NOT NULL,
+  `email`      VARCHAR(255) NOT NULL,
+  `user_type`  SMALLINT     NOT NULL,
+  `group_id`   BIGINT       NULL,
+  `access_key` VARCHAR(100) NOT NULL,
   PRIMARY KEY (`id`),
   FOREIGN KEY (group_id)
-    REFERENCES student_groups(id)
-    ON UPDATE RESTRICT ON DELETE CASCADE
-) DEFAULT CHARSET=utf8;
+  REFERENCES student_groups(id)
+    ON UPDATE RESTRICT
+    ON DELETE CASCADE
+)
+  DEFAULT CHARSET = utf8;
